backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pymongo==4.6.1
dnspython==2.4.2
python-dotenv==1.0.0
django-cors-headers==4.3.1
bcrypt==4.1.2
